Flask_SQLAlchemy_HTMx_Bookstore/pyproject.toml
[tool.poetry]
name = "flask-sqlalchemy-htmx-bookstore"
version = "0.1.0"
description = ""
authors = ["Mike Jackson <perigran_falcon@yahoo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.1.1"
flask-sqlalchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
